Support for $translate plugin
New attribute slide-tabs-translatable (true/false), this attribute works with plugin angular-translate, so instead of putting an ion-slide-tab-label="Title here" it supports the translate key -> ion-slide-tab-label="KEY.title". and the directive will render the $translate.instant('KEY.title');
Also the directive is watching for any changes in default display language so it will re-render the KEY.title with new output.